Add files via upload
Support for Exodia based around "The Legendary Exodia Incarnate" and it's new spin for the deck. The art for 3 of the cards are from official Yu-Gi-Oh! media, and I got permission to use the art for Exodia the Empowered from Andy Tantowi, the creator of the art as long as I credited him. From my testing the cards seem to preform rather fairly, though the thought to change Exodia the Empowered's effect so that it targets instead has crossed my mind. While this does give the deck some new powerful plays, it's still open to disruption.